feature/file-storage/refactor: deleted pkg/storage -> added service/minio
