+++ internal/app/app.go
@@ -3,6 +3,7 @@ package app
 import (
 	"context"
 	"errors"
+	"github.com/zenorachi/dynamic-user-segmentation/internal/service/minio"
 	"net/http"
 	"os"
 	"os/signal"
@@ -51,6 +52,7 @@ func Run(cfg *config.Config) {
 		TokenManager:    tokenManager,
 		AccessTokenTTL:  cfg.Auth.AccessTokenTTL,
 		RefreshTokenTTL: cfg.Auth.RefreshTokenTTL,
+		Storage:         minio.NewProvider(&cfg.Minio),
 	})
 
 	/* INIT HTTP HANDLER */
